% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument.R
\name{apply_left.Argument}
\alias{apply_left.Argument}
\title{"Argument" apply_left.}
\usage{
\method{apply_left}{Argument}(pipe_left_arg, pipe_right_arg,
  pipe_environment, left_arg_name, pipe_string, right_arg_name)
}
\arguments{
\item{pipe_left_arg}{left argument}

\item{pipe_right_arg}{substitute(pipe_right_arg) argument}

\item{pipe_environment}{environment to evaluate in}

\item{left_arg_name}{name, if not NULL name of left argument.}

\item{pipe_string}{character, name of pipe operator.}

\item{right_arg_name}{name, if not NULL name of right argument.}
}
\value{
result
}
\description{
\code{apply_left} for \code{Argument} \code{class}.  Forced evaluation of
right argument and appies \code{apply_right}.
}
\examples{

as_argument(1:3) \%.>\% wrapfn(sin, "x") \%.>\% wrapfn(cos, "x") \%.>\% as_value(.)

}
