% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CapturePipeline.R
\name{apply_left.Collector}
\alias{apply_left.Collector}
\title{Capture all right-arguments unevaluated.}
\usage{
\method{apply_left}{Collector}(pipe_left_arg, pipe_right_arg,
  pipe_environment, left_arg_name, pipe_string, right_arg_name)
}
\arguments{
\item{pipe_left_arg}{left argument}

\item{pipe_right_arg}{substitute(pipe_right_arg) argument}

\item{pipe_environment}{environment to evaluate in}

\item{left_arg_name}{name, if not NULL name of left argument.}

\item{pipe_string}{character, name of pipe operator.}

\item{right_arg_name}{name, if not NULL name of right argument.}
}
\value{
result
}
\description{
Capture all right-arguments unevaluated, using \code{bquote()-.()} escaping.
}
\keyword{internal}
