% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpipe.R
\name{apply_left}
\alias{apply_left}
\title{S3 dispatch on class of pipe_left_arg.}
\usage{
apply_left(
  pipe_left_arg,
  pipe_right_arg,
  pipe_environment,
  left_arg_name,
  pipe_string,
  right_arg_name
)
}
\arguments{
\item{pipe_left_arg}{left argument.}

\item{pipe_right_arg}{substitute(pipe_right_arg) argument.}

\item{pipe_environment}{environment to evaluate in.}

\item{left_arg_name}{name, if not NULL name of left argument.}

\item{pipe_string}{character, name of pipe operator.}

\item{right_arg_name}{name, if not NULL name of right argument.}
}
\value{
result
}
\description{
For formal documentation please see \url{https://github.com/WinVector/wrapr/blob/master/extras/wrapr_pipe.pdf}.
}
\examples{

apply_left.character <- function(pipe_left_arg,
                                 pipe_right_arg,
                                 pipe_environment,
                                 left_arg_name,
                                 pipe_string,
                                 right_arg_name) {
  if(is.language(pipe_right_arg)) {
    wrapr::apply_left_default(pipe_left_arg,
                              pipe_right_arg,
                              pipe_environment,
                              left_arg_name,
                              pipe_string,
                              right_arg_name)
  } else {
    paste(pipe_left_arg, pipe_right_arg)
  }
}
setMethod(
  wrapr::apply_right_S4,
  signature = c(pipe_left_arg = "character", pipe_right_arg = "character"),
  function(pipe_left_arg,
           pipe_right_arg,
           pipe_environment,
           left_arg_name,
           pipe_string,
           right_arg_name) {
    paste(pipe_left_arg, pipe_right_arg)
  })

"a" \%.>\% 5 \%.>\% 7

"a" \%.>\% toupper(.)

q <- "z"
"a" \%.>\% q


}
\seealso{
\code{\link{apply_left.default}}
}
