% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnaryFunctions.R
\name{apply_left.UnaryFn}
\alias{apply_left.UnaryFn}
\title{Apply right wrapped function to argument on left.}
\usage{
\method{apply_left}{UnaryFn}(pipe_left_arg, pipe_right_arg,
  pipe_environment, left_arg_name, pipe_string, right_arg_name)
}
\arguments{
\item{pipe_left_arg}{left argument should be a class derived from UnaryFn.}

\item{pipe_right_arg}{substitute(pipe_right_arg) argument, should evaluate to a class derived from UnaryFn.}

\item{pipe_environment}{environment to evaluate in}

\item{left_arg_name}{name, if not NULL name of left argument.}

\item{pipe_string}{character, name of pipe operator.}

\item{right_arg_name}{name, if not NULL name of right argument.}
}
\value{
result
}
\description{
Apply right wrapped function to argument on left.
}
\keyword{internal}
