% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpipe.R
\name{apply_left_default}
\alias{apply_left_default}
\title{S3 dispatch on class of pipe_left_arg.}
\usage{
apply_left_default(pipe_left_arg, pipe_right_arg, pipe_environment,
  left_arg_name, pipe_string, right_arg_name)
}
\arguments{
\item{pipe_left_arg}{left argument}

\item{pipe_right_arg}{substitute(pipe_right_arg) argument}

\item{pipe_environment}{environment to evaluate in}

\item{left_arg_name}{name, if not NULL name of left argument.}

\item{pipe_string}{character, name of pipe operator.}

\item{right_arg_name}{name, if not NULL name of right argument.}
}
\value{
result
}
\description{
Place evaluation of left argument in \code{.} and then evaluate right argument.
}
\examples{

5 \%.>\% sin(.)

}
\seealso{
\code{\link{apply_left}}
}
