% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpipe.R
\name{apply_right.default}
\alias{apply_right.default}
\title{Default apply_right implementation.}
\usage{
\method{apply_right}{default}(pipe_left_arg, pipe_right_arg,
  pipe_environment, left_arg_name, pipe_string, right_arg_name)
}
\arguments{
\item{pipe_left_arg}{left argument}

\item{pipe_right_arg}{pipe_right_arg argument}

\item{pipe_environment}{environment to evaluate in}

\item{left_arg_name}{name, if not NULL name of left argument.}

\item{pipe_string}{character, name of pipe operator.}

\item{right_arg_name}{name, if not NULL name of right argument.}
}
\value{
result
}
\description{
Default apply_right implementation: S4 dispatch to apply_right_S4.
}
\examples{

# simulate a function pointer
apply_right.list <- function(pipe_left_arg,
                             pipe_right_arg,
                             pipe_environment,
                             left_arg_name,
                             pipe_string,
                             right_arg_name) {
  pipe_right_arg$f(pipe_left_arg)
}

f <- list(f=sin)
2 \%.>\% f
f$f <- cos
2 \%.>\% f

}
\seealso{
\code{\link{apply_left}}, \code{\link{apply_right}}, \code{\link{apply_right_S4}}
}
