% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locum.R
\name{apply_right.locum}
\alias{apply_right.locum}
\title{S3 dispatch on class of pipe_right_argument for a locum.}
\usage{
\method{apply_right}{locum}(pipe_left_arg, pipe_right_arg,
  pipe_environment, left_arg_name, pipe_string, right_arg_name)
}
\arguments{
\item{pipe_left_arg}{left argument}

\item{pipe_right_arg}{right argument}

\item{pipe_environment}{environment to evaluate in}

\item{left_arg_name}{name, if not NULL name of left argument.}

\item{pipe_string}{character, name of pipe operator.}

\item{right_arg_name}{name, if not NULL name of right argument.}
}
\value{
result
}
\description{
Triggered if right hand side of pipe stage was a name that does not resolve to a function.
For formal documentation please see \url{https://github.com/WinVector/wrapr/blob/master/extras/wrapr_pipe.pdf}.
}
