% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funobj.R
\name{apply_right.pipe_list}
\alias{apply_right.pipe_list}
\title{Apply right to a list of items.}
\usage{
\method{apply_right}{pipe_list}(pipe_left_arg, pipe_right_arg,
  pipe_environment, left_arg_name, pipe_string, right_arg_name)
}
\arguments{
\item{pipe_left_arg}{left argument}

\item{pipe_right_arg}{a pipe_list}

\item{pipe_environment}{environment to evaluate in}

\item{left_arg_name}{name, if not NULL name of left argument.}

\item{pipe_string}{character, name of pipe operator.}

\item{right_arg_name}{name, if not NULL name of right argument.}
}
\value{
result
}
\description{
Apply right to a list of items.
}
\keyword{internal}
