% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funobj.R
\name{applyto}
\alias{applyto}
\title{Apply a wrapped function to an argument.}
\usage{
applyto(wfn, arg, env = parent.frame())
}
\arguments{
\item{wfn}{wrapped function.}

\item{arg}{additional argument value.}

\item{env}{environment to evaluate in.}
}
\value{
wfn applied to arg.
}
\description{
Apply a wrapped function to an argument.
}
\seealso{
\code{\link{wrap_function_S3}}, \code{\link{wrap_fname_S3}}, \code{\link{wrap_function_S4}}, \code{\link{wrap_fname_S4}}
}
