% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnaryFunctions.R
\name{as.list.UnaryFnList}
\alias{as.list.UnaryFnList}
\title{Get list of primative unary fns.}
\usage{
\method{as.list}{UnaryFnList}(x, ...)
}
\arguments{
\item{x}{UnaryFn derived classe to extract}

\item{...}{not used.}
}
\value{
list of non UnaryFnList functions
}
\description{
Get list of primative unary fns.
}
\examples{

as.list(pkgfn("base::sin", "x"))
as.list(c(pkgfn("base::sin", "x"), pkgfn("base::cos", "x")))

}
