% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument.R
\name{as_argument.Argument}
\alias{as_argument.Argument}
\title{Argument constructor}
\usage{
\method{as_argument}{Argument}(x)
}
\arguments{
\item{x}{object to wrap}
}
\value{
x wrapped in an Argument container.
}
\description{
Pass through unaltered.
}
