% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument.R
\name{as_argument.default}
\alias{as_argument.default}
\title{Argument constructor}
\usage{
\method{as_argument}{default}(x)
}
\arguments{
\item{x}{object to wrap}
}
\value{
x wrapped in an Argument container.
}
\description{
Build an Argument class from an object.
}
