% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument.R
\name{as_argument}
\alias{as_argument}
\title{Argument constructor}
\usage{
as_argument(x)
}
\arguments{
\item{x}{object to wrap}
}
\value{
x wrapped in an Argument container.
}
\description{
Build an Argument class from an object.
}
