% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dot_fn.R
\name{as_fn}
\alias{as_fn}
\title{Convert a pipeable object into a function.}
\usage{
as_fn(pipeable, env = parent.frame())
}
\arguments{
\item{pipeable}{a wrapr dot-pipe pipeable object}

\item{env}{envirnonment to work in.}
}
\value{
single function with signature (., env = parent.frame())
}
\description{
Convert a pipeable object into a function of "."
}
\details{
Note: writes "." into env.
}
\examples{

p <- pkgfn("base::sin", "x")
f <- as_fn(p)
f(5)


}
