% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnaryFunctions.R
\name{as_fnlist}
\alias{as_fnlist}
\title{Wrap a list of functions as a function.}
\usage{
as_fnlist(items)
}
\arguments{
\item{items}{list of UnaryFn derived instances (no nested UnaryFnList items).}
}
\value{
UnaryFnList
}
\description{
Wrap a list of functions as a function.
}
\examples{

f <- as_fnlist(list(pkgfn("base::sin", "x"), pkgfn("base::cos", "x")))
cat(format(f))
1:3 \%.>\% f

}
\seealso{
\code{\link{pkgfn}}, \code{\link{wrapfn}}, \code{\link{srcfn}}
}
