% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_named_list.R
\name{as_named_list}
\alias{as_named_list}
\title{Capture named objects as a named list.}
\usage{
as_named_list(...)
}
\arguments{
\item{...}{argument names (must be names, not strings or values).}
}
\value{
a named list mapping argument names to argument values
}
\description{
Build a named list from a sequence of named argumetn of the form NAME = VALUE.
}
\examples{

a <- data.frame(x = 1)
b <- 2
str(as_named_list(a, b))

# an example application for this function is managing saving and
# loading values into the workspace.
if(FALSE) {
  library(wrapr)

  # example values
  a <- 5
  b <- 7
  do_not_want <- 13

  # save the elements of our workspace we want
  saveRDS(as_named_list(a, b), 'example_data.RDS')

  # clear values out of our workspace for the example
  rm(list = ls())
  ls()

  # read back while documenting what we expect to
  # read in
  readRDS('example_data.RDS') -> unpack[a, b]

  # confirm what we have, the extra unpack is a side
  # effect of the -> notation. To avoid this write
  # instead:
  # readRDS('example_data.RDS') \%.>\% unpack[a, b]
  ls()

  print(a)

  print(b)
}

}
