% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument.R
\name{as_value.Argument}
\alias{as_value.Argument}
\title{Argument value (extract).}
\usage{
\method{as_value}{Argument}(x)
}
\arguments{
\item{x}{Argument class}
}
\value{
x[[1]]
}
\description{
Argument value (extract).
}
