% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument.R
\name{as_value.default}
\alias{as_value.default}
\title{Default Argument value (self).}
\usage{
\method{as_value}{default}(x)
}
\arguments{
\item{x}{object}
}
\value{
x
}
\description{
Default Argument value (self).
}
