% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument.R
\name{as_value}
\alias{as_value}
\title{Convert from Argument to value.}
\usage{
as_value(x)
}
\arguments{
\item{x}{Argument or general object}
}
\value{
x object
}
\description{
Convert from Argument to value.
}
