% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atblock.R
\name{ateval}
\alias{ateval}
\title{Evaluate an expression with \code{@name} substitution (deprecated).}
\usage{
ateval(.)
}
\arguments{
\item{.}{character text of expression or block to evaluate}
}
\description{
The expression represented by the text of the \code{.}-argument is evaluated in
the current environment will all "\code{@name}" forms replaced by the value of name.
This allows variables to carry the names of other variables into what would
be a non-standard evaluation situation.  This adaption allows some parametric
or standard-evaluation effects in such circumstances.
}
\details{
Note: this method uses string substitution and is willing to substitute in arbitrary content,
please prefer using \code{\link{let}} where applicable.

Notation idea: Jonathan Carroll \url{https://jcarroll.com.au} \url{https://twitter.com/carroll_jono/status/842142292253196290}
Similar to: \url{https://dev.mysql.com/doc/refman/5.7/en/user-variables.html}.
}
\seealso{
\code{\link{let}}
}
