% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_reg_choro.R
\name{bc_reg_choro}
\alias{bc_reg_choro}
\title{Make a choropleth map of BC economic regions}
\usage{
bc_reg_choro(tbbl, region, thingy, value, num_format)
}
\arguments{
\item{tbbl}{a tibble that contains (at least) the following 3 columns:}

\item{region}{the column of the tbbl that contains the economic regions: these MUST be camel_case i.e. mainland_south_west, vancouver_island_coast, north_coast_&_nechako, cariboo, kootenay, north_east, thompson_okanagan}

\item{thingy}{the column of the tbbl that contains the name of what is being plotted.}

\item{value}{the column of the tbble that contains the value of what is being plotted.}

\item{num_format}{whether the number being plotted should be formatted as a comma or percent}
}
\description{
Make a choropleth map of BC economic regions
}
