% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_reg_pop.R
\name{bc_reg_pop}
\alias{bc_reg_pop}
\title{Gets the most recent measure of population for BCs economic geographic_areas}
\usage{
bc_reg_pop()
}
\value{
tibble
}
\description{
Gets the most recent measure of population for BCs economic geographic_areas
}
