% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_reg_sf.R
\name{bc_reg_sf}
\alias{bc_reg_sf}
\title{Gets a simple features dataframe for BCs economic regions (for mapping)}
\usage{
bc_reg_sf()
}
\value{
sf tibble
}
\description{
Gets a simple features dataframe for BCs economic regions (for mapping)
}
