% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atblock.R
\name{beval}
\alias{beval}
\title{Evaluate an expression with (!!name) substitution.}
\usage{
beval(...)
}
\arguments{
\item{...}{expression or block to evaluate}
}
\description{
The expression represented by the text of the .-argument is evaluated in
the current environment will all "(!!name)" forms replaced by the value of name.
Note: this substitution can only be performed on the right hand side of assignments.
This allows variables to carry the names of other variables into what would
be a non-standard evaluation situation.  This adaption allows some parametric
or standard-evaluation effects in such circumstances.
}
\details{
Notation idea: \url{https://github.com/hadley/dplyr/commit/8f03f835185370626a566e95d268623b20189e07},
}
\examples{


c1 <- 1:5
c2 <- 3:7
col1 <- 'c1'
col2 <- 'c2'
beval(
  res <- (!!col1) + (!!col2)
)
print(res)


}
