% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bquotefn.R
\name{bquote_call}
\alias{bquote_call}
\title{Treat ... call argument as bquoted-values.}
\usage{
bquote_call(call, env = parent.frame())
}
\arguments{
\item{call}{result of match.call()}

\item{env}{environment to perform lookups in.}
}
\value{
altered call
}
\description{
bquote_call re-writes calls.
}
\details{
Note: eagerly evalutes argument and writes them into the function's
executing environment.
}
\seealso{
\code{\link{bquote_function}}, \code{\link{bquote_call_args}}
}
\keyword{internal}
