% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bquotefn.R
\name{bquote_function}
\alias{bquote_function}
\title{Adapt a function to use bquote on its arguments.}
\usage{
bquote_function(fn)
}
\arguments{
\item{fn}{function to adapt, must have non-empty formals().}
}
\value{
new function.
}
\description{
Adapt a function to use bquote on its arguments.
}
\examples{

f1 <- function(x) { substitute(x) }
f2 <- bquote_function(f1)
arg <- as.name("USER_ARG")
f2(arg)    # returns arg
f2(.(arg)) # returns USER_ARG

}
