% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnaryFunctions.R
\name{c.UnaryFn}
\alias{c.UnaryFn}
\title{Combine UnaryFns}
\usage{
\method{c}{UnaryFn}(...)
}
\arguments{
\item{...}{UnaryFn derived classes to combine}
}
\value{
UnaryFn representing the sequence
}
\description{
Combine UnaryFns
}
\examples{

c(pkgfn("base::sin", "x"), pkgfn("base::cos", "x"))

}
