% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camel_to_title.R
\name{camel_to_title}
\alias{camel_to_title}
\title{convert all factors in a tibble to Title Case character/}
\usage{
camel_to_title(tbbl)
}
\arguments{
\item{tbbl}{a tibble}
}
\value{
a tibble
}
\description{
convert all factors in a tibble to Title Case character/
}
