% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_frames.R
\name{check_equiv_frames}
\alias{check_equiv_frames}
\title{Check two data.frames are equivalent after sorting columns and rows.}
\usage{
check_equiv_frames(d1, d2, ..., tolerance = sqrt(.Machine$double.eps))
}
\arguments{
\item{d1}{data.frame 1}

\item{d2}{data.frame 2}

\item{...}{force later arguments to bind by name}

\item{tolerance}{numeric comparision tolerance}
}
\value{
logical TRUE if equivalent
}
\description{
Confirm two dataframes are equivalent after reordering columns and rows.
}
