% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fit.R
\name{clean_fit_lm}
\alias{clean_fit_lm}
\title{Fit a stats::lm without carying back large structures.}
\usage{
clean_fit_lm(outcome, variables, data, ..., weights = NULL,
  env = baseenv())
}
\arguments{
\item{outcome}{character, name of outcome column.}

\item{variables}{character, names of varaible columns.}

\item{data}{data.frame, training data.}

\item{...}{not used, force later arguments to be used by name}

\item{weights}{passed to stats::glm()}

\item{env}{environment to work in.}
}
\value{
list(model=model, summary=summary)
}
\description{
Fit a stats::lm without carying back large structures.
}
\examples{

k = 1000
dlarge <- data.frame(
  x1 = rep(c("a", "a", "b", "b"), k),
  x2 = rep(c(0, 0, 0, 1), k),
  y = rep(1:4, k),
  yC = rep(c(FALSE, TRUE, TRUE, TRUE), k),
  stringsAsFactors = FALSE)
dsmall <- data.frame(
  x1 = c("a", "a", "b", "b"),
  x2 = c(0, 0, 0, 1),
  y = 1:4,
  yC = c(FALSE, TRUE, TRUE, TRUE),
  stringsAsFactors = FALSE)
res_lm <- clean_fit_lm("y", c("x1", "x2"), dlarge)
str(res_lm$model)
predict(res_lm$model, newdata = dsmall)

}
