% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tbbl.R
\name{clean_tbbl}
\alias{clean_tbbl}
\title{Wrapper function that calls clean_names and converts character columns to clean factors}
\usage{
clean_tbbl(tbbl)
}
\arguments{
\item{tbbl}{a tibble}
}
\value{
a clean tibble
}
\description{
Wrapper function that calls clean_names and converts character columns to clean factors
}
