% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CapturePipeline.R
\name{Collector}
\alias{Collector}
\title{Build a collector that can capture all pipe stages to the right.}
\usage{
Collector()
}
\value{
a Collector list-object.
}
\description{
Build a collector that can capture all pipe stages to the right, using
\code{bquote()-.()} escaping.
}
\examples{

phase <- 0.1
Collector() \%.>\% sin(.) \%.>\% cos(. + .(phase))

}
