% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnaryFunctions.R
\name{concat_items_rev}
\alias{concat_items_rev}
\title{build a list of all UnaryFn from possibly composite}
\usage{
concat_items_rev(op1, op2)
}
\arguments{
\item{op1}{list1}

\item{op2}{list2}
}
\value{
c(list2, list1)
}
\description{
build a list of all UnaryFn from possibly composite
}
\keyword{internal}
