% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funobj.R
\name{def_funobj_s4_class}
\alias{def_funobj_s4_class}
\title{Define a S4 right action class.}
\usage{
def_funobj_s4_class(Class, where = topenv(parent.frame()))
}
\arguments{
\item{Class}{character, name of class.}

\item{where}{environment to work in.}
}
\description{
Define a S4 right action class.
}
