% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{define_unpacker}
\alias{define_unpacker}
\title{Create a value unpacking object.}
\usage{
define_unpacker(object_name)
}
\arguments{
\item{object_name}{character, name the object is stored as}
}
\value{
an unpacker
}
\description{
Create a value unplacking object that records it is stored by name \code{object_name}.
}
\keyword{internal}
