% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut.R
\name{donut}
\alias{donut}
\title{Donut plot}
\usage{
donut(
  df,
  y_var,
  group_var,
  center_text = NULL,
  center_text_size = 7,
  caption = NULL,
  lab_size = 5
)
}
\arguments{
\item{df}{dataframe that contains}

\item{y_var}{The measure being plotted}

\item{group_var}{a grouping variable}

\item{center_text}{what text to put in middle, Default: NULL}

\item{center_text_size}{how large center font is, Default: 7}

\item{caption}{user provided caption, Default: NULL}

\item{lab_size}{font size for labels, Default: 5}
}
\value{
a donut chart
}
\description{
takes a dataframe with a grouping variable and a measure and makes a donut.
}
