% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_clutch.R
\name{doubleapply}
\alias{doubleapply}
\alias{\%!\%}
\title{Double apply pipe.}
\usage{
pipe_left_arg \%!\% pipe_right_arg
}
\arguments{
\item{pipe_left_arg}{left argument expression (substituted into .)}

\item{pipe_right_arg}{right argument expession (presumably including .)}
}
\value{
\code{pipe_left_arg \%.>\% (pipe_left_arg \%.>\% pipe_right_arg)}
}
\description{
Defined as roughly : \code{a \%!\% b} ~ \code{a \%.>\% (a \%.>\% b)}.
Useful for systems like \code{rquery} where first appearance of an object
captures a description and second appearance applies the operation.
}
\examples{

# An example function that captures a desciption on
# first pass and value on the second.
# The meaninful application of this is rquery pipelines.
b <- function(a1) {
  a_name <- substitute(a1)
  function(a2) {
    paste0(a_name, ": ", a2)
  }
}

x <- 7

x \%!\% b # "x: 7"
5 \%!\% b # "5: 5"

}
