% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/let.R
\name{echoAndExecute}
\alias{echoAndExecute}
\title{Conditionally echo and always execute expression (equivilant to let() with no alias).}
\usage{
echoAndExecute(expr, ..., debugPrint = FALSE)
}
\arguments{
\item{expr}{block to prepare for execution.}

\item{...}{force later arguments to be bound by name.}

\item{debugPrint}{logical if TRUE print debugging information}
}
\value{
result of expr executed in calling environment
}
\description{
Conditionally echo and always execute expression (equivilant to let() with no alias).
}
\examples{

echoAndExecute(1+1, debugPrint= FALSE)
echoAndExecute(1+1, debugPrint= TRUE)

}
