% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CapturePipeline.R
\name{eval_dot_sequence}
\alias{eval_dot_sequence}
\title{Evaluate a sequence of expressions with \code{.}-value substituted in.}
\usage{
eval_dot_sequence(., dot_seq, env = parent.frame())
}
\arguments{
\item{.}{value to pass in.}

\item{dot_seq}{list of expressions.}

\item{env}{environment to work in.}
}
\value{
sequential evaluation result.
}
\description{
Evaluate a sequence of expressions with \code{.}-value substituted in.
}
\details{
Note: not for steps that intend side-effects or have references to items
in non-standard environments.
}
\keyword{internal}
