% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_labs.R
\name{fix_labs}
\alias{fix_labs}
\title{fix labels on a ggplot2 object}
\usage{
fix_labs(gg)
}
\arguments{
\item{gg}{a ggplot2 object}
}
\value{
a ggplot2 object
}
\description{
fix labels on a ggplot2 object
}
\seealso{
\code{\link[ggplot2]{labs}}
\code{\link[stringr]{case}},\code{\link[stringr]{str_replace}}
}
