% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnaryFunctions.R
\name{fnlist}
\alias{fnlist}
\title{Wrap a list of functions as a function.}
\usage{
fnlist(items)
}
\arguments{
\item{items}{list of UnaryFn derived instances (no nested UnaryFnList items).}
}
\value{
UnaryFnList
}
\description{
Wrap a list of functions as a function.
}
\seealso{
\code{\link{pkgfn}}, \code{\link{wrapfn}}, \code{\link{srcfn}}
}
