% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locum.R
\name{format.locum}
\alias{format.locum}
\title{Format a locum for presentation.}
\usage{
\method{format}{locum}(x, ...)
}
\arguments{
\item{x}{locum to be formatted}

\item{...}{additional arguments, use "start" to replace initial step presentation}
}
\value{
formatted string
}
\description{
Format a locum for presentation.
}
\examples{

p <- locum() \%.>\% sin(.)
format(p, start = 5)

}
