% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnaryFunctions.R
\name{format.PartialFunction}
\alias{format.PartialFunction}
\title{format step}
\usage{
\method{format}{PartialFunction}(x, ...)
}
\arguments{
\item{x}{object to format}

\item{...}{additional aguments (not used)}
}
\value{
character
}
\description{
format step
}
