% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpipe.R
\name{\%.>\%}
\alias{\%.>\%}
\title{Pipe-like operator ("dot block pipe").}
\usage{
a \%.>\% b
}
\arguments{
\item{a}{left argument (substituted into .)}

\item{b}{right argument (presumably including .)}
}
\description{
Defined as: a %.>% b  ~roughly~  { . <- a; b }; (though, we try to stop .-side effects).
(please see \url{http://www.win-vector.com/blog/2017/07/in-praise-of-syntactic-sugar/}).
}
\examples{

cos(exp(sin(4)))
4 \%.>\% sin(.) \%.>\% exp(.) \%.>\% cos(.)

}
