% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_interpolation.R
\name{\%s\%}
\alias{\%s\%}
\title{Dot substitution string interpolation.}
\usage{
str \%s\% envir
}
\arguments{
\item{str}{charater string to be substituted into}

\item{envir}{environemnt to look for values}
}
\value{
modified strings
}
\description{
String interpolation using \code{bquote}-stype .() notation. Pure R, no C/C++ code called.
}
\details{
See also
\url{https://CRAN.R-project.org/package=R.utils},
\url{https://CRAN.R-project.org/package=rprintf},
and \url{https://CRAN.R-project.org/package=glue}.
}
\examples{

"x is .(x), x+1 is .(x+1)\n.(x) is odd is .(x\%\%2 == 1)" \%s\% list(x = 7)


}
\seealso{
\code{\link{strsplit_capture}}, \code{\link{si}}
}
