% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_interpolation.R
\name{\%s>\%}
\alias{\%s>\%}
\title{Dot substitution string interpolation.}
\usage{
envir \%s>\% str
}
\arguments{
\item{envir}{environemnt to look for values}

\item{str}{charater string to be substituted into}
}
\value{
modified strings
}
\description{
String interpolation using \code{bquote}-stype .() notation. Pure R, no C/C++ code called.
}
\details{
See also
\url{https://CRAN.R-project.org/package=R.utils},
\url{https://CRAN.R-project.org/package=rprintf},
and \url{https://CRAN.R-project.org/package=glue}.
}
\examples{

list(x = 7) \%s>\% "x is .(x)"


}
\seealso{
\code{\link{strsplit_capture}}, \code{\link{si}}
}
