% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepv.R
\name{grepv}
\alias{grepv}
\title{Return a vector of matches.}
\usage{
grepv(
  pattern,
  x,
  ...,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE,
  invert = FALSE
)
}
\arguments{
\item{pattern}{character scalar, pattern to match, passed to \code{\link[base]{grep}}.}

\item{x}{character vector to match to, passed to \code{\link[base]{grep}}.}

\item{...}{not used, forced later arguments to bind by name.}

\item{ignore.case}{logical, passed to \code{\link[base]{grep}}.}

\item{perl}{logical, passed to \code{\link[base]{grep}}.}

\item{fixed}{logical, passed to \code{\link[base]{grep}}.}

\item{useBytes}{logical, passed \code{\link[base]{grep}}.}

\item{invert}{passed to \code{\link[base]{grep}}.}
}
\value{
vector of matching values.
}
\description{
Return a vector of matches.
}
\examples{

grepv("x$", c("sox", "xor"))

}
\seealso{
\code{\link[base]{grep}}, \code{\link{grepdf}}
}
