% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_frames.R
\name{has_no_dup_rows}
\alias{has_no_dup_rows}
\title{Check for duplicate rows.}
\usage{
has_no_dup_rows(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
TRUE if there are no duplicate rows, else FALSE.
}
\description{
Check a simple data.frame (no list or exotic rows) for duplicate rows.
}
