% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{\%c\%}
\alias{\%c\%}
\title{Inline list/array concatenate.}
\usage{
e1 \%c\% e2
}
\arguments{
\item{e1}{first, or left argument.}

\item{e2}{second, or right argument.}
}
\value{
c(e1, c2)
}
\description{
Inline list/array concatenate.
}
\examples{

1:2 \%c\% 5:6

c("a", "b") \%c\% "d"

}
