% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{\%dot\%}
\alias{\%dot\%}
\title{Inline dot product.}
\usage{
e1 \%dot\% e2
}
\arguments{
\item{e1}{first, or left argument.}

\item{e2}{second, or right argument.}
}
\value{
c(e1, c2)
}
\description{
Inline dot product.
}
\examples{

c(1,2) \%dot\% c(3, 5)


}
