% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{\%+\%}
\alias{\%+\%}
\title{Inline character concatenate.}
\usage{
e1 \%+\% e2
}
\arguments{
\item{e1}{first, or left argument.}

\item{e2}{second, or right argument.}
}
\value{
c(e1, c2)
}
\description{
Inline character concatenate.
}
\examples{

"a" \%+\% "b"

c("a", "b") \%+\% "_d"

}
