% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{\%qc\%}
\alias{\%qc\%}
\title{Inline quoting list/array concatenate.}
\usage{
e1 \%qc\% e2
}
\arguments{
\item{e1}{first, or left argument.}

\item{e2}{second, or right argument.}
}
\value{
qc(e1, c2)
}
\description{
Inline quoting list/array concatenate.
}
\examples{

1:2 \%qc\% 5:6

c("a", "b") \%qc\% d

a \%qc\% b \%qc\% c

}
