% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\docType{data}
\name{into}
\alias{into}
\title{Unpack or bind values into the calling environment.}
\format{An object of class \code{unpacker} of length 1.}
\usage{
into
}
\description{
Unpacks or binds values into the calling environment. Uses \code{bquote} escaping.
NULL is a special case that is unpacked to all targets. NA targets are skipped.
All non-NA target names must be unique.
}
\details{
Similar to \code{Python} tuple unpacking, \code{zeallot}'s arrow, and to \code{vadr::bind}.

Note: a reference to the unpacker object is written into the unpacking environment as a side-effect
of the implied array assignment.  Also, can not unpack into a variable name same as the un-packer's
original declared name (into).
}
\examples{

# name capture version
into[a, b] <- list(5, 10)
print(a)  # now 5
print(b)  # now 10

# bquote re-direct to value in variable using .()
# Note: the bquote .() step is potentially confusing, as the user
# can't immediately see where the value is being assigned to.
# Also, quotes are allowed.
a <- 'x'
into[.(a), 'b'] <- list(20, 40)
print(x)  # now 20
print(b)  # now 40
print(a)  # still 'x'

}
\keyword{datasets}
