% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_order.R
\name{invert_perm}
\alias{invert_perm}
\title{Invert a permuation.}
\usage{
invert_perm(p)
}
\arguments{
\item{p}{vector of lenth n containing each of seq_len(n) exactly once.}
}
\value{
vector q such that p[q] == q[p] == seq_len(length(p))
}
\description{
For a permutation p build q such that p[q] == q[p] == seq_len(length(p)).
See \url{http://www.win-vector.com/blog/2017/05/on-indexing-operators-and-composition/}.
}
\examples{

p <- c(4, 5, 7, 8, 9, 6, 1, 3, 2, 10)
q <- invert_perm(p)
p[q]
all.equal(p[q], seq_len(length(p)))
q[p]
all.equal(q[p], seq_len(length(p)))

}
