% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda}
\alias{lambda}
\title{Build an anonymous function.}
\usage{
lambda(..., env = parent.frame())
}
\arguments{
\item{...}{formal parameters of function, unbound names, followed by function body (code/language).}

\item{env}{environment to work in}
}
\value{
user defined function.
}
\description{
Mostly just a place-holder so lambda-symbol form has somewhere safe to hang its help entry.
}
\examples{

#lambda-syntax: lambda(arg [, arg]*, body [, env=env])
# also works with lambda character as function name
# print(intToUtf8(0x03BB))

# example: square numbers
sapply(1:4, lambda(x, x^2))

# example more than one argument
f <- lambda(x, y, x+y)
f(2,4)


}
\seealso{
\code{\link{defineLambda}}, \code{\link{makeFunction_se}}, \code{\link{named_map_builder}}
}
