% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorizeM.R
\name{lapplym}
\alias{lapplym}
\title{Memoizing wrapper for lapply.}
\usage{
lapplym(X, FUN, ...)
}
\arguments{
\item{X}{list or vector of inputs}

\item{FUN}{function to apply}

\item{...}{additional arguments passed to lapply}
}
\value{
list of results.
}
\description{
Memoizing wrapper for lapply.
}
\examples{

fs <- function(x) { x <- x[[1]]; print(paste("see", x)); sin(x) }
# should only print "see" twice, not 6 times
lapplym(c(0, 1, 1, 0, 0, 1), fs)

}
\seealso{
\code{\link{VectorizeM}}, \code{\link{vapplym}}, \code{\link{parLapplyLBm}}
}
