% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funobj.R
\name{lapplyto}
\alias{lapplyto}
\title{Apply a list of wrapped functions to an argument.}
\usage{
lapplyto(lwfn, arg, env = parent.frame())
}
\arguments{
\item{lwfn}{list of wrapped functions}

\item{arg}{additional argument value.}

\item{env}{environment to evaluate in.}
}
\value{
wfn applied to arg.
}
\description{
Apply a list of wrapped functions to an argument.
}
\seealso{
\code{\link{wrap_function_S3}}, \code{\link{wrap_fname_S3}}, \code{\link{wrap_function_S4}}, \code{\link{wrap_fname_S4}}
}
