% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/let.R
\name{letprep_str}
\alias{letprep_str}
\title{Substitute text (note text can be a vector).}
\usage{
letprep_str(alias, strexpr, ...)
}
\arguments{
\item{alias}{mapping named list/vector to strings/names or general}

\item{strexpr}{character vector source text to be re-written}

\item{...}{force later arguments to be bound by name.}
}
\value{
parsed R expression with substitutions
}
\description{
Substitute text (note text can be a vector).
}
