% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/let.R
\name{letprepl}
\alias{letprepl}
\title{Substitute language elements.}
\usage{
letprepl(alias, lexpr)
}
\arguments{
\item{alias}{mapping named list/vector to strings/names or general}

\item{lexpr}{language item}

\item{...}{force later arguments to be bound by name.}
}
\value{
R language element with substitutions
}
\description{
Substitute language elements.
}
