% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locum.R
\name{locum}
\alias{locum}
\title{Build a stand in for a future value to be placed in a pipe.}
\usage{
locum()
}
\value{
a locum stand-in
}
\description{
The locum stands in for a value to be specified later in a pipeline.
This is similar to a lambda or function abstraction.
}
\examples{

p <- locum() \%.>\% sin(.)
5 \%.>\% p

}
