% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clean_factor.R
\name{make_clean_factor}
\alias{make_clean_factor}
\title{dirty string to factor with camel_case levels.}
\usage{
make_clean_factor(strng)
}
\arguments{
\item{strng}{a string}
}
\value{
a factor
}
\description{
dirty string to factor with camel_case levels.
}
