% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_title.R
\name{make_title}
\alias{make_title}
\title{converts a camel_case string to Title Case.}
\usage{
make_title(strng)
}
\arguments{
\item{strng}{a camel_case string}
}
\value{
a Title Case string
}
\description{
converts a camel_case string to Title Case.
}
