% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvars.R
\name{mapvars}
\alias{mapvars}
\title{Map varaible names to referenced values if those values are string scalars (else throw).}
\usage{
mapvars(...)
}
\arguments{
\item{...}{varaible names mapping to string scalars}
}
\value{
map from original variable names to new names
}
\description{
Map varaible names to referenced values if those values are string scalars (else throw).
}
\examples{

x <- 'a'
y <- 'b'
mapvars(x, y)
d <- data.frame(a = 1, b = 2)
let(mapvars(x, y), d$x + d$y)

}
