% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnaryFunctions.R
\docType{class}
\name{PartialFunction-class}
\alias{PartialFunction-class}
\title{Function with partial arguments as a new single argument function.}
\description{
Function with partial arguments as a new single argument function.
}
