% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funobj.R
\name{pipe_list}
\alias{pipe_list}
\title{Construct a list of class pipe_list}
\usage{
pipe_list(...)
}
\arguments{
\item{...}{items to keep}
}
\value{
list of class pipe_list
}
\description{
Construct a list of class pipe_list
}
