% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnaryFunctions.R
\name{pkgfn}
\alias{pkgfn}
\title{Wrap the name of a function as a function.}
\usage{
pkgfn(fname, arg_name = ".", args = list())
}
\arguments{
\item{fname}{character, function name in fname or package::fname format.}

\item{arg_name}{characer, name of argument to assign.}

\item{args}{named list of adittional arguments and values.}
}
\value{
PartialNamedFn
}
\description{
Wrap the name of a function as a function.
}
\examples{

f <- pkgfn("base::sin", "x")
cat(format(f))
1:3 \%.>\% f

}
\seealso{
\code{\link{fnlist}}, \code{\link{wrapfn}}, \code{\link{srcfn}}
}
