% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locum.R
\name{print.locum}
\alias{print.locum}
\title{Print a locum presentation.}
\usage{
\method{print}{locum}(x, ...)
}
\arguments{
\item{x}{locum to be formatted}

\item{...}{additional arguments, use "start" to replace initial step presentation}
}
\value{
formatted string
}
\description{
Print a locum presentation.
}
\examples{

p <- locum() \%.>\% sin(.)
print(p, start = 5)

}
