% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc}
\alias{qc}
\title{Quoting version of c() array concatinator.}
\usage{
qc(...)
}
\arguments{
\item{...}{items to place into an array}
}
\value{
quoted array of character items
}
\description{
Note: evaluation of qc() arguments happens in the qc() environment, not the calling environment.
This means assigments inside a qc() argument may not be visible, also name collisions with
\code{...} or \code{.wrapr_private_var_} prefixed variables are possible.
}
\examples{

qc(a, qc(b, c))
qc(x=a, qc(y=b, z=c))
qc('x'='a', qc('y'='b', 'z'='c'))

}
\seealso{
\code{\link{qe}}, \code{\link{qae}}
}
