% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/let.R
\name{restrictToNameAssignments}
\alias{restrictToNameAssignments}
\title{Restrict an alias mapping list to things that look like name assignments}
\usage{
restrictToNameAssignments(alias, restrictToAllCaps = TRUE)
}
\arguments{
\item{alias}{mapping list}

\item{restrictToAllCaps}{logical, if true only use all-capitalized keys}
}
\value{
string to string mapping
}
\description{
Restrict an alias mapping list to things that look like name assignments
}
\examples{

alias <- list(region= 'east', str= "'seven'")
aliasR <- restrictToNameAssignments(alias)
print(aliasR)


}

