% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_package_tests.R
\name{run_package_tests}
\alias{run_package_tests}
\title{Run package tests.}
\usage{
run_package_tests(pkg, ..., verbose = TRUE, test_dir = NULL,
  stop_on_issue = TRUE, stop_if_no_tests = TRUE)
}
\arguments{
\item{pkg}{character, name of package to test.}

\item{...}{not used, force later arguments to bind by name.}

\item{verbose}{logical, if TRUE print more.}

\item{test_dir}{directory to look for tests in, if not set looks in pacakge unit_tests.}

\item{stop_on_issue}{logical, if TRUE stop after errors or failures.}

\item{stop_if_no_tests}{logical, if TRUE stop if no tests were found.}
}
\value{
nothing
}
\description{
For all files with names of the form "^test_.+\\.R$" in the package directory unit_tests
run all functions with names of the form "^test_.+$" as RUnit tests.
Attaches RUnit and pkg, requires RUnit.
Stops on error.
}
\details{
Based on \url{https://github.com/RcppCore/Rcpp/blob/master/tests/doRUnit.R}.
}
