% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_fns.R
\name{sel_rows}
\alias{sel_rows}
\title{Pick a selection of rows (can be used to order).}
\usage{
sel_rows(x, rows)
}
\arguments{
\item{x}{data.frame to work with}

\item{rows}{numeric, row indexes.}
}
\value{
data.frame that is the specified row selection.
}
\description{
Only works on in-memory data.frames.
Part piping with base R series: \url{http://www.win-vector.com/blog/tag/piping-with-base-r/}.
}
\examples{

d <- data.frame(x = c('b', 'a', 'c'))
sel_rows(d, d$x)

}
