% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_fns.R
\name{select_columns}
\alias{select_columns}
\title{Pick a subset of columns.}
\usage{
select_columns(x, columns)
}
\arguments{
\item{x}{data.frame to work with}

\item{columns}{character, names of columns}
}
\value{
data.frame that is the specified subset of the columns of x.
}
\description{
Only works on in-memory data.frames.
Part piping with base R series: \url{http://www.win-vector.com/blog/tag/piping-with-base-r/}.
}
\examples{

head(select_columns(mtcars, c("mpg", "cyl", "disp")))

}
