% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CapturePipeline.R
\name{sequence_as_function}
\alias{sequence_as_function}
\title{Convert a sequence of expressions into a function.}
\usage{
sequence_as_function(dot_seq, env = parent.frame())
}
\arguments{
\item{dot_seq}{list of expressions.}

\item{env}{environment to work in.}
}
\value{
function with signature (., eval_environment = parent.frame())
}
\description{
Convert a sequence of expressions into a function.
}
\details{
Note: not for steps that intend side-effects or have references to items
in non-standard environments.
}
\examples{

seq <- Collector() \%.>\% paste(., "a") \%.>\% paste(., "b")
f <- sequence_as_function(seq)
f("x")

}
