% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funobj.R
\name{set_funobj_s4_applyto}
\alias{set_funobj_s4_applyto}
\title{Set the S4 right action.}
\usage{
set_funobj_s4_applyto(left_class, right_class,
  where = topenv(parent.frame()))
}
\arguments{
\item{left_class}{character, name of left class.}

\item{right_class}{character, name of right class.}

\item{where}{environment to work in.}
}
\description{
Set the S4 right action of apply_right_S4 for left_class and right_class.
}
\keyword{internal}
