% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funobj.R
\name{set_funobj_s4_right_action}
\alias{set_funobj_s4_right_action}
\title{Set the S4 right action of apply_right_S4.}
\usage{
set_funobj_s4_right_action(left_class, where = topenv(parent.frame()))
}
\arguments{
\item{left_class}{character, name of left class.}
}
\description{
Set the S4 right action of apply_right_S4 for left_class and "wrapr_funobj_S4".
}
