% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnaryFunctions.R
\docType{methods}
\name{show,UnaryFnList-method}
\alias{show,UnaryFnList-method}
\title{S4 print method}
\usage{
\S4method{show}{UnaryFnList}(object)
}
\arguments{
\item{object}{item to print}
}
\description{
S4 print method
}
