% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_at_brace_pairs.R
\name{split_at_brace_pairs}
\alias{split_at_brace_pairs}
\title{Split strings at {}-pairs.}
\usage{
split_at_brace_pairs(s)
}
\arguments{
\item{s}{string or list of strings to split.}
}
\value{
array or list of split strings.
}
\description{
Split strings at {}-pairs.
}
\examples{

split_at_brace_pairs("{x} + y + {z}")

}
