% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{[.unpacker}
\alias{[.unpacker}
\title{Build an unpackign target.}
\usage{
\method{[}{unpacker}(self, ...)
}
\arguments{
\item{self}{object implementing the feature, wrapr::unpack}

\item{...}{names of to unpack to (can be escaped with bquote \code{.()} notation).}
}
\value{
self
}
\description{
Unpacks or binds values into the calling environment. Uses \code{bquote} escaping.
NULL is a special case that is unpacked to all targets. NA targets are skipped.
All non-NA target names must be unique.
}
\details{
Similar to \code{Python} tuple unpacking, \code{zeallot}'s arrow, and to \code{vadr::bind}.
}
\examples{

# into and to have the attribute dotpipe_eager_eval set to TRUE,
# so they do not need to be written as .(intro(a, b)) to trigger
# eager function eval in the pipeline
list(5, 10) \%.>\% into[a, b]
print(a)  # now 5
print(b)  # now 10

}
