% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{[<-.unpacker}
\alias{[<-.unpacker}
\title{Unpack or bind values into the calling environment.}
\usage{
\method{[}{unpacker}(self, ...) <- value
}
\arguments{
\item{self}{object implementing the feature, wrapr::unpack}

\item{...}{names of to unpack to (can be escaped with bquote \code{.()} notation).}

\item{value}{list to unpack into values, must have a number of entries equal to number of \code{...} arguments}
}
\value{
self
}
\description{
Unpacks or binds values into the calling environment. Uses \code{bquote} escaping.
}
\details{
Similar to \code{Python} tuple unpacking, \code{zeallot}'s arrow, and to \code{vadr::bind}.
}
\examples{

# name capture version
into[a, b] <- list(5, 10)
print(a)  # now 5
print(b)  # now 10

# bquote re-direct to value in variable using .()
# Note: the bquote .() step is potentially confusing, as the user
# can't immediately see where the value is being assigned to.
# Also, quotes are allowed.
a <- 'x'
into[.(a), 'b'] <- list(20, 40)
print(x)  # now 20
print(b)  # now 40
print(a)  # still 'x'

}
