% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\docType{data}
\name{unpack}
\alias{unpack}
\title{Unpack or bind values into the calling environment.}
\format{An object of class \code{unpacker} of length 0.}
\usage{
unpack
}
\description{
Unpacks or binds values into the calling environment. Uses bquote escaping.
}
\details{
Similar to \code{Python} tuple unpacking, \code{zeallot}'s arrow, and to \code{vadr::bind}.
}
\examples{

# name capture version
unpack[a, b] <- list(5, 10)
print(a)  # now 5
print(b)  # now 10

# bquote re-direct to value in variable using .()
# plus quotes are allowed
a <- 'x'
unpack[.(a), 'b'] <- list(20, 40)
print(x)  # now 20
print(b)  # now 40
print(a)  # still 'x'

}
\keyword{datasets}
