% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{unpack_to}
\alias{unpack_to}
\title{Unpack values into current environment.}
\usage{
unpack_to(value, ...)
}
\arguments{
\item{value}{list of values}

\item{...}{names to unpack to}
}
\description{
Unpack the values in the \code{value} argument to names given in the \code{...} argument
}
\examples{

list(7, 12) \%.>\% unpack_to(., a, b)
print(a)  # now 7
print(b)  # now 12

}
