% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{Unpacker}
\alias{Unpacker}
\title{Create a value unpacking object.}
\usage{
Unpacker(object_name = NULL, allow_unnamed_case = FALSE)
}
\arguments{
\item{object_name}{character, name the object is stored as}

\item{allow_unnamed_case}{logical, if FALSE promote unnamed arguments to self-named}
}
\value{
an unpacker
}
\description{
Create a value unplacking object that records it is stored by name \code{object_name}.
}
\details{
Array-assign form can not use the names: \code{.}, \code{wrapr_private_self}, \code{value}, or the name stored in \code{object_name}.
function form can not use the names: \code{.} or \code{wrapr_private_value}. Array-form will wrong own name into working environment
as a side-effect.
}
\keyword{internal}
