% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{UnpackerF}
\alias{UnpackerF}
\title{Create a value unpacking object (function version).}
\usage{
UnpackerF(object_name = NULL)
}
\arguments{
\item{object_name}{character, name the object is stored as}
}
\value{
an unpacker
}
\description{
Create a value unplacking object that records it is stored by name \code{object_name} (function version).
}
\details{
Array-assign form can not use the names: \code{.}, \code{wrapr_private_self}, \code{value}, or the name stored in \code{object_name}.
function form can not use the names: \code{.} or \code{wrapr_private_value}. Array-form will wrong own name into working environment
as a side-effect.
}
\keyword{internal}
