% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{UnpackerP}
\alias{UnpackerP}
\title{Create a value unpacking object (eager pipe version).}
\usage{
UnpackerP(object_name = NULL)
}
\arguments{
\item{object_name}{character, name the object is stored as}
}
\value{
an unpacker
}
\description{
Create a value unplacking object that records it is stored by name \code{object_name} (eager pipe version).
}
\details{
Array-assign form can not use the names: \code{.}, \code{wrapr_private_self}, \code{value}, or the name stored in \code{object_name}.
function form can not use the names: \code{.} or \code{wrapr_private_value}.
Array-form with \code{=}, \code{<-}, \code{->} will write own name into working environment
as a side-effect. Array-form with \code{:=} does not have the side-effect.
}
\keyword{internal}
