% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{UnpackerSE}
\alias{UnpackerSE}
\title{Create a value unpacking object (standard evaluation variant).}
\usage{
UnpackerSE(object_name = NULL, unnamed_case = FALSE)
}
\arguments{
\item{object_name}{character, name the object is stored as}

\item{unnamed_case}{logical, if TRUE use positional- not name based matching}
}
\value{
an unpacker
}
\description{
Create a value unplacking object that records it is stored by name \code{object_name}.
This is included merely for completeness.
}
\keyword{internal}
