% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{Invoke a spreadsheet like viewer when appropriate.}
\usage{
view(x, ..., title = wrapr_deparse(substitute(x)), n = 200)
}
\arguments{
\item{x}{R object to view}

\item{...}{force later arguments to bind by name.}

\item{title}{title for viewer}

\item{n}{number of rows to show}
}
\value{
invoke view or format object
}
\description{
Invoke a spreadsheet like viewer when appropriate.
}
\examples{

view(mtcars)

}
