% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funobj.R
\name{wrap_fname_S3}
\alias{wrap_fname_S3}
\title{Wrap a function name and arguments in an object for later S3 pipe dispatch.}
\usage{
wrap_fname_S3(fn_name = NULL, ..., fn_package = "base",
  arg_name = "", args = list())
}
\arguments{
\item{fn_name}{character, name of function.}

\item{...}{force later arguments to be taken by name.}

\item{fn_package}{character, name of package.}

\item{arg_name}{name for remaining argument.}

\item{args}{list of function argument values}
}
\value{
wrapped function
}
\description{
Wrap a function name and arguments in an object for later S3 pipe dispatch.
}
\seealso{
\code{\link{applyto}}
}
