% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funobj.R
\name{wrap_function_S4}
\alias{wrap_function_S4}
\title{Wrap a function and arguments in an object for later S4 pipe dispatch.}
\usage{
wrap_function_S4(fn, ..., arg_name = "", args = list())
}
\arguments{
\item{fn}{function to wrap}

\item{...}{force later arguments to be taken by name.}

\item{arg_name}{name for remaining argument.}

\item{args}{list of function argument values}
}
\value{
wrapped function
}
\description{
Wrap a function and arguments in an object for later S4 pipe dispatch.
}
\seealso{
\code{\link{applyto}}
}
