% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnaryFunctions.R
\name{wrapfn}
\alias{wrapfn}
\title{Wrap the source for an exprssion as a function.}
\usage{
wrapfn(fn, arg_name = ".", args = list())
}
\arguments{
\item{fn}{function.}

\item{arg_name}{characer, name of argument to assign.}

\item{args}{named list of adittional arguments and values.}
}
\value{
PartialFunction
}
\description{
Wrap the source for an exprssion as a function.
}
\examples{

f <- wrapfn(sin, "x")
cat(format(f))
1:3 \%.>\% f

}
\seealso{
\code{\link{pkgfn}}, \code{\link{fnlist}}, \code{\link{srcfn}}
}
