% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{wrapr}
\alias{wrapr}
\alias{wrapr-package}
\title{wrapr: Wrap R Functions for Debugging and Ease of Use}
\description{
Methods to capture function context on error for debugging.
Also methods to convert non-standard evaluation interfaces to
parameterized standard evaluation interfaces.
}
\details{
For more information:
\itemize{
  \item \code{vignette('wrapr', package='wrapr')}
  \item \code{vignette(package='wrapr')}
  \item Website: \url{https://github.com/WinVector/wrapr} }
}

