% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpipe.R
\name{wrapr_function.default}
\alias{wrapr_function.default}
\title{Wrapr function.}
\usage{
\method{wrapr_function}{default}(pipe_left_arg, pipe_right_arg,
  pipe_environment, pipe_name = NULL)
}
\arguments{
\item{pipe_left_arg}{left argument.}

\item{pipe_right_arg}{right argument (general object, not a function).}

\item{pipe_environment}{environment to evaluate in.}

\item{pipe_name}{character, name of pipe operator.}
}
\value{
result
}
\description{
S3 dispatch on tyhpe of pipe_right_argument.
}
\examples{

f <- function() { print("execute"); 0}
a <- substitute({. + 1 + f()})
5 \%.>\% a


}
