% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OldPipeFns.R
\name{wrapr_function}
\alias{wrapr_function}
\title{wrapr_function}
\usage{
wrapr_function(pipe_left_arg, pipe_right_arg, pipe_environment,
  pipe_name = NULL)
}
\arguments{
\item{pipe_left_arg}{left argument.}

\item{pipe_right_arg}{right argument (general object, not a function).}

\item{pipe_environment}{environment to evaluate in.}

\item{pipe_name}{character, name of pipe operator.}
}
\value{
result
}
\description{
Out of date, please use \code{\link{apply_right}} instead.
}
