% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{write_values_into_env}
\alias{write_values_into_env}
\title{write values into environment}
\usage{
write_values_into_env(unpack_environment, str_args, value)
}
\arguments{
\item{unpack_environment}{environment to write into}

\item{str_args}{array of string-names to write to}

\item{value}{values to write}
}
\description{
write values into environment
}
\keyword{internal}
