% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc.R
\name{bc}
\alias{bc}
\title{Parse argument as a vector of values allowing "blank separators".}
\usage{
bc(s)
}
\arguments{
\item{s}{string to parse}
}
\value{
vector of values
}
\description{
Accepts simple string that do not include escapes. Separates data on whitespace, comma, and pipe.
Pattern matching adapted for US English.
Suggested by Emil Erik Pula Bellamy Begtrup-Bright \url{https://github.com/WinVector/wrapr/issues/12}.
}
\examples{

bc('1 2 "c", d') # returns c("1", "2", "c", "d")
bc('1 2 3') # returns c(1, 2, 3)
bc('1 2 "3"') # returns c("1", "2", "3")

}
