% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atblock.R
\name{beval}
\alias{beval}
\title{Evaluate an expression with \code{(!!name)} and \code{:=} to \code{=} substitution (deprecated).}
\usage{
beval(...)
}
\arguments{
\item{...}{expression or block to evaluate}
}
\description{
Notation idea: \url{https://github.com/hadley/dplyr/commit/8f03f835185370626a566e95d268623b20189e07}.
Note: "\code{!!}" is not a no-op, but is a sufficiently uncommon expression one can use it.
}
\seealso{
\code{\link{let}}
}
