% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bquotefn.R
\name{bquote_call}
\alias{bquote_call}
\title{Treat call argument as bquoted-values.}
\usage{
bquote_call(call, env = parent.frame())
}
\arguments{
\item{call}{result of match.call()}

\item{env}{environment to perform lookups in.}
}
\value{
altered call
}
\description{
Re-write call to evaluate \code{expr} with \code{bquote} \code{.()} substitution.
Uses convetion that := is considered a alias for =.
Including \code{.(-x)} promoting \code{x}'s value from character to a name,
which is called "quote negation" (hence the minus-sign).
}
\seealso{
\code{\link{bquote_function}}, \code{\link{bquote_call_args}}
}
\keyword{internal}
