% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bquotefn.R
\name{bquote_function}
\alias{bquote_function}
\title{Adapt a function to use bquote on its arguments.}
\usage{
bquote_function(fn)
}
\arguments{
\item{fn}{function to adapt, must have non-empty formals().}
}
\value{
new function.
}
\description{
bquote_function is for adapting a function defined elsewhere for bquite-enabled argument substitution.
}
\examples{



# angle = 1:10
# variable <- as.name("angle")
# plotb <- bquote_function(graphics::plot)
# plotb(x = .(variable), y = sin(.(variable)))



f1 <- function(x) { substitute(x) }
f2 <- bquote_function(f1)
arg <- as.name("USER_ARG")
f2(arg)    # returns arg
f2(.(arg)) # returns USER_ARG




}
\seealso{
\code{\link{bquote_call_args}}
}
