% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{defineLambda}
\alias{defineLambda}
\title{Define lambda function building function.}
\usage{
defineLambda(envir = parent.frame(), name = NULL)
}
\arguments{
\item{envir}{environment to work in.}

\item{name}{character, name to assign to (defaults to Greek lambda).}
}
\description{
Use this to place a copy of the lambda-symbol
function builder in your workspace.
}
\examples{

defineLambda()
# ls()

}
