% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bquotefn.R
\name{evalb}
\alias{evalb}
\title{eval(bquote(expr)) shortcut.}
\usage{
evalb(..., where = parent.frame())
}
\arguments{
\item{...}{expression to evaluate (one argument).}

\item{where}{environment to work in.}
}
\value{
eval(bquote(expr))
}
\description{
eval(bquote(expr)) shortcut.
}
\examples{

angle = 1:10
variable <- as.name("angle")
evalb(plot(x = .(variable), y = sin(.(variable))))

}
