% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnaryFunctions.R
\name{fnlist}
\alias{fnlist}
\title{Wrap a list of functions as a function.}
\usage{
fnlist(...)
}
\arguments{
\item{...}{UnaryFn derived instances.}
}
\value{
UnaryFnList
}
\description{
Unary functions are evaluated in left to right or first to last order.
}
\examples{

f <- fnlist(pkgfn("base::sin", "x"), pkgfn("base::cos", "x"))
cat(format(f))
1:3 \%.>\% f

}
\seealso{
\code{\link{pkgfn}}, \code{\link{wrapfn}}, \code{\link{srcfn}}
}
