% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpipe.R
\name{\%.>\%}
\alias{\%.>\%}
\title{Pipe operator ("dot arrow").}
\usage{
pipe_left_arg \%.>\% pipe_right_arg
}
\arguments{
\item{pipe_left_arg}{left argument expression (substituted into .)}

\item{pipe_right_arg}{right argument expession (presumably including .)}
}
\value{
eval(\{ . <- pipe_left_arg; pipe_right_arg \};)
}
\description{
Defined as: \code{a \%.>\% b} roughly ~ \code{\{ . <- a; b \};}
(with visible .-side effects).
Please see \url{http://www.win-vector.com/blog/2017/07/in-praise-of-syntactic-sugar/}.
}
\examples{

# both should be equal:
cos(exp(sin(4)))
4 \%.>\% sin(.) \%.>\% exp(.) \%.>\% cos(.)

}
