% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{\%p\%}
\alias{\%p\%}
\title{Inline character paste0.}
\usage{
e1 \%p\% e2
}
\arguments{
\item{e1}{first, or left argument.}

\item{e2}{second, or right argument.}
}
\value{
c(e1, c2)
}
\description{
Inline character paste0.
}
\examples{

"a" \%p\% "b"

c("a", "b") \%p\% "_d"

}
