% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/let.R
\name{letprep_str}
\alias{letprep_str}
\title{Substitute text.}
\usage{
letprep_str(alias, strexpr, ..., debugPrint = FALSE)
}
\arguments{
\item{alias}{mapping named list/vector to strings/names or general}

\item{strexpr}{character vector source text to be re-written}

\item{...}{force later arguments to be bound by name.}

\item{debugPrint}{logical if TRUE print debugging information}
}
\value{
parsed R expression with substitutions
}
\description{
Substitute text.
}
