% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_char.R
\name{map_to_char}
\alias{map_to_char}
\title{format a map.}
\usage{
map_to_char(mp, sep = "\\n  ")
}
\arguments{
\item{mp}{named vector or list}

\item{sep}{separator}
}
\value{
character formatted representation
}
\description{
format a map.
}
\examples{

cat(map_to_char(c('a':='b', 'c':='d')))

}
