% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpipe.R
\name{pipe_impl}
\alias{pipe_impl}
\title{Pipe dispatch implementation.}
\usage{
pipe_impl(pipe_left_arg, pipe_right_arg, pipe_environment, pipe_string = NULL)
}
\arguments{
\item{pipe_left_arg}{substitute(pipe_left_arg) argument.}

\item{pipe_right_arg}{substitute(pipe_right_arg) argument.}

\item{pipe_environment}{environment to evaluate in.}

\item{pipe_string}{character, name of pipe operator.}
}
\value{
result
}
\description{
This is a helper for implementing additional pipes.
}
\keyword{internal}
