% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpipe.R
\name{pipe_step}
\alias{pipe_step}
\title{Pipe step operator}
\usage{
pipe_step(pipe_left_arg, pipe_right_arg, pipe_environment, pipe_name = NULL)
}
\arguments{
\item{pipe_left_arg}{left argument.}

\item{pipe_right_arg}{substitute(pipe_right_arg) argument.}

\item{pipe_environment}{environment to evaluate in.}

\item{pipe_name}{character, name of pipe operator.}
}
\value{
result
}
\description{
Pipe step operator
}
