% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qae.R
\name{qs}
\alias{qs}
\title{Quote argument as a string.}
\usage{
qs(s)
}
\arguments{
\item{s}{expression to be quoted as a string.}
}
\value{
character
}
\description{
qs() uses bquote() .() quasiquotation escaping notation.
}
\examples{

x <- 7

qs(a == x)

qs(a == .(x))

}
