% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{[.Unpacker}
\alias{[.Unpacker}
\title{Prepare for unpack or bind values into the calling environment.}
\usage{
\method{[}{Unpacker}(wrapr_private_self, ...)
}
\arguments{
\item{wrapr_private_self}{object implementing the feature, wrapr::unpack}

\item{...}{names of to unpack to (can be escaped with bquote \code{.()} notation).}
}
\value{
prepared unpacking object
}
\description{
Prepare for unpack or bind values into the calling environment.  This makes pipe to behavior very
close to assign to behavior for the Unpacker class.
}
