% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{unpacker_target}
\alias{unpacker_target}
\title{create an argname carrier that is a function}
\usage{
unpacker_target(unpack_environment, ...)
}
\arguments{
\item{unpack_environment}{environment to look into}

\item{...}{names to capture for assignments}
}
\value{
an unpacking function
}
\description{
create an argname carrier that is a function
}
\keyword{internal}
