% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{wrapr}
\alias{wrapr}
\alias{wrapr-package}
\title{\code{wrapr}: Wrap R Functions for Debugging and Ease of Use}
\description{
Provides \code{DebugFnW} to capture function context on error for
debugging, and \code{let} which converts non-standard evaluation interfaces to
parameterized standard evaluation interfaces.
}
\details{
For more information:
\itemize{
  \item \code{vignette('DebugFnW', package='wrapr')}
  \item \code{vignette('let', package='wrapr')}
  \item \code{vignette(package='wrapr')}
  \item Website: \url{https://github.com/WinVector/wrapr}
  \item \code{let} video: \url{https://youtu.be/iKLGxzzm9Hk?list=PLAKBwakacHbQp_Z66asDnjn-0qttTO-o9}
  \item Debug wrapper video: \url{https://youtu.be/zFEC9-1XSN8?list=PLAKBwakacHbQT51nPHex1on3YNCCmggZA}.}
}

