% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpipe.R
\name{wrapr_function}
\alias{wrapr_function}
\title{Wrapr function.}
\usage{
wrapr_function(pipe_left_arg, pipe_right_arg, pipe_environment,
  pipe_name = NULL)
}
\arguments{
\item{pipe_left_arg}{left argument.}

\item{pipe_right_arg}{right argument.}

\item{pipe_environment}{environment to evaluate in.}

\item{pipe_name}{character, name of pipe operator.}
}
\value{
result
}
\description{
S3 dispatch on tyhpe of pipe_right_argument.
}
